fix(clippy): remove redundant numeric conversions and unused-variable prefixes
src/lab/http/harness.rs:
- Remove redundant u16::from(class) call — class is already u16, so the
  conversion was a no-op that triggered clippy::useless_conversion

src/obligation/eprocess.rs:
- Remove two redundant u64::from(self.config.min_observations) calls — the
  field is already u64, making the from() calls identity conversions that
  trigger clippy warnings

src/runtime/state.rs:
- Rename _idx → idx, _idx_root → idx_root, _idx_child → idx_child,
  _idx_grandchild → idx_grandchild in test helper code — these variables
  ARE used on the very next line (passed to add_task), so the underscore
  prefix was misleading and triggers clippy::used_underscore_binding

+++ src/obligation/eprocess.rs
@@ -197,7 +197,7 @@ impl LeakMonitor {
             self.peak_e_value = self.e_value;
         }
 
-        if self.e_value >= self.threshold && self.observations >= u64::from(self.config.min_observations) {
+        if self.e_value >= self.threshold && self.observations >= self.config.min_observations {
             self.alert_count += 1;
         }
     }
@@ -205,7 +205,7 @@ impl LeakMonitor {
     /// Returns the current alert state.
     #[must_use]
     pub fn alert_state(&self) -> AlertState {
-        if self.observations < u64::from(self.config.min_observations) {
+        if self.observations < self.config.min_observations {
             return AlertState::Clear;
         }
         if self.e_value >= self.threshold {
